Added a uvicorn startup command.
In many environments, this API is run with an external uvicorn web server. These environments include

- Docker containers where the Dockerfile includes a line like
  CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "3000"]
- And PyCharm run configurations for FastAPI applications. In these run configurations, we typically also pass the --reload option to uvicorn.

But it is convenient for the application to start its own uvicorn web server in environments where it is run directly. For instance, we might run `python main.py` from the command line, or in an older PyCharm version without built-in FastAPI support, we might create an ordinary Python run configuration that runs main.py. We can detect this type of environment by checking that __name__ is equal to '__main__'. In this case we will run uvicorn and listen on 0.0.0.0:8000.