Meta-analysis score sets
We add API support for setting and reading source score sets of meta-analyses. In the view model for creating a score set, these are specified as a list of source score set URNs. View models for reading score sets include source score sets; the ShortScoreset view used in lists contexts is an exception.

Source score sets cannot currently be set when updating a score set. This could easily be allowed, but perhaps it would complicate usage.