Rename `pubmed_identifier` Model to `publication_identifier`
Future support for preprints and other publictions means that we will no
longer just be accepting pubmed identifiers as publications. This change
renames the `pubmed_identifier` model to `publication_identifier`, so
that it will be more clear in the future this model generalizes to all
our supported publications. At the moment, we still only support pubmed
publications.