Add Support for Primary and Secondary Publications (Experiments)
Adds support for primary and secondary publications when creating and
editing experiments. Whether a publication is defined as `primary` is
enumerated in the `experiment_publication_identifiers` association
table. This required changing the previous relationship object to an
association_proxy object. This allowed us to define a new
`ExperimentPublicationIdentifierAssociation` class that was able to
add an extra field to the table, `primary`, that stores whether a
publication is a primary one for a given experiment.

Because this `association_proxy` object returns an internal and untyped
`AssociationList` object, a new validator was added to our view model
that could properly handle this object. We now return an additional
object to our UI layer: `ExperimentPublicationIdentifierAssociations`.
This object is used by the UI to differentiate between primary and
secondary publications.

Note that when we receive primary publications through our experiments
router, we add a temporary `primary` attribute to them. This allows
our association proxy object to store the correct value for `primary`.