Add Support for Primary and Secondary Publications (Scoresets)
Adds support for primary and secondary publications when creating and
editing scoresets. Whether or not a publication is defined as `primary`
is enumerated in the `scoreset_publication_identifiers` association
table. We now have two relationships acting on this table. One is an
AssociationProxy relationship defined by a new type,
`ScoresetPublicationIdentifierAssociation`. The other acts on our
already defined `PublicationIdentifier` objects.

When a scoreset is posted to its router, these association objects handle
the associations between scoreset and publication identifiers. They also
add whether an associated publication is the `primary` one for a given
scoreset or not. Although UI restrictions currently will only post one
primary publication, there is no reason that a scoreset couldn't have
multiple primary publications if this functionality is needed in the
future.