Endpoint renaming (#68)
Use kebab case in API URLs, and treat "score set" as two words.

We inherited the treatment of 'scoreset' as a single word from the Django application, but the revisions here treat it as two words, except in the underlying database table and column names. (These can be refactored later.)

In addition, multi-word path elements in API endpoint URLs are revised here to use kebab casing consistently.

In particular:
- API endpoint paths involving 'scoreset' now use 'score-set.'
- Model class properties and view model properties involving 'scoreset' now use 'score_set,' which is exposed in APIs as 'scoreSet.'
- Text involving 'scoreset' now uses 'score set.'
- All API endpoint path elements now use kebab case.
- The only remaining occurrence of 'scoreset' as a single word is in database table and column names.