Re-Validate Variants on Score Set Update
Previous versions of this logic flow allowed users to create a temp scoreset with
valid variant/target mappings, then edit the temp score set and in-validate
variant/target mappings by editing the target. This change re-validates variants
when users engage in this workflow and discards edits to the score set if new
variants are invalid.