Add Support for Custom Validation Locations
There are some cases where we would like to perform view model
validation at a level higher in the model than where the error might be.
In these cases, it is useful to provide custom error locations to our
validation error so that the proper loc is surfaced in the API error
response. This can be extremely useful information to the user and for
creating meaningful validation error support in UI elements.