Add Label Field to Target Sequence Model
Adds a label field to the target sequence model. This label field will
be used to define variants with respect to a specific target sequence in
cases where multiple target genes are defined on one score set. Note
that this label should be alphanumeric but may be omitted so long as
only one target gene is defined on a score set.