Optimize SQL queries for score set searches.
Calls to search_score_sets have been resulting in a large number of queries, often lazy-loading individual relationships separately for each row in the search results. Here we use SQLAlchemy's eager-loading options to reduce the number of queries.

With this change, calls to search_score_sets involve
- One query that loads matching score sets, together with all of the relationships that will be returned in the ShortScoreSet view model;
- And a series of subqueries for all the score sets belonging to the experiments from the first query, together with all of the necessary relationships.

The subqueries use SELECT IN to keep the number of queries small -- currently, three queries cover all published experiments.

We might revisit whether all of the relationships in ShortScoreSet are really necessary for search results.