Add CDOT package and samtools/htslib Installation to Docker Image
Adds the CDOT package to our requirements. This will allow us to expand the pool of variants we
accept for genomic coordinate score-sets. Adds samtools and htslib to our Docker image, along
with downloading data files CDOT requires for transcript resolution. Samtools and htslib are both
required to zip and index these data files. Data for both the GRCh37 and GRCh38 builds is included
with this change.

Uses a multi-stage build for the CDOT Docker data so that our final image only includes the required
data files and not the tools that build them. These tools are not needed by our final product, so we
don't need to include them in the image.