Use CDOT to Resolve Transcripts when not Present in UTA
Uses the built-in `ChainedSeqFetcher` provided with CDOT to resolve transcripts that are potentially
missing from UTA. This is surfaced as a new Dep to ensure that all instances of our genomic coordinate
validators and seqfetchers use the same data pool. If a transcript is present in UTA, we will prefer
that sequence. If a transcript is not present in UTA, we will use the transcripts associated with the
human genome builds GRCh38 and GRCh37.