Bugfix: If one target was not of type `dna`, validation of `hgvs_nt` fails
Fixes a bug where if one target was not of type `dna` in a multi target upload, the validation of column
`hgvs_nt` would fail. This occured because targets were checked individually on whether they satisfied a
columns' requirements. Now, we ensure that at least one target is of type `dna` if a scoreset contains the
`hgvs_nt` column. If we attempt to validate a protein variant within this column, it will fail later during
variant validation.

This change also clarifies some thrown errors during this column validation step, providing more useful and
actionable output if targets fail validation for a column.