Add Statistics Router
Creates a new router endpoint, `/statistics`, containing routes which calculate counting stats for
certain fields on Target Genes, Experiments, and Score Sets. Endpoints are validated against enumerated
fields that denote which model properties are eligible for calculations. To add a new field, add the field
name to the appropriate model field list and add any custom logic for statistic calculation to the helper
functions within the router definition. Built in FastAPI methods validate requests to this router to ensure
any non-enumerated fields are responded to with the appropriate error.