Use HGVS Data Provider Directly on HGVS Routes
Transitions to using the HGVS dependency data provider on most HGVS routes (the one exception is
the `/genes` route, of which UTA provides a more complete list). This change eliminates all
references to and usage of SeqRepo based methods within the HGVS router.

This change also alters router logic so that empty or NoneType responses that would have surfaced a
200 response + an empty data field now respond with a 404 error. Tests have been edited to reflect
this change.