Process Variants in Async Worker Process
Adds worker job for processing workers asynchronously. To process variants for a score set,
enqueue this job to the worker with the score set urn, editing user id, scores, and counts.
The job will process these scores/counts and either insert variants into the database and
mark the score set processing state as successful or fail to validate/process the variants
and mark the score set processing state as failed.

Generally, internal errors are not saved to the database (although score sets are still marked
as failed) while errors stemming from user data- usually ValidationErrors- are inserted to
the database so that the user may remedy them.