Add Trigger List to ValidationError
Adds a list of triggers to the ValidationError class. This makes it easier to
handle ValidationErrors where multiple individual HGVS strings were invalid, as
they can be surfaced as a list of errors rather than a string joined via commas.
This change does not alter how other places where this error is raised interact
with it.