Fix TargetGeneCreate Validator for Optional Target Subtypes
Fixes the TargetGeneCreate validator for optional target accession and target sequence subtypes. The
previous validator was intented to ensure that both `target_sequence` and `target_accession` weren't
defined on a single TargetGene instance, but was not validating every TargetGene (since `always` was
not set to `True`).

This change makes the validator a `root_validator` so that it triggers every time, and also checks to
ensure that one of either `target_sequence` or `target_accession` is defined on an instance of a
target gene in addition to checking that they are mutually exclusive. A couple new tests are included
to test this new behavior (and to ensure any reversions to the prior, buggy, behavior are caught).

Co-authored-by: Estelle Da <8445097+estelleda@users.noreply.github.com>