Use server-side database streaming (cursors) when preparing variant CSV strings. (#91)
By refraining from holding all variants in memory, we improve the speed of all variant CSV generation. This change is particularly necessary for the full database dump because otherwise, for certain large score sets, the SQLAlchemy Variant model objects fill up available memory and cause the application to be killed.

Notice that, in variant CSV exports, we now order variants numerically by their URN fragment. The code for this is PostgreSQL-specific and slows the query; we could add a "variant number" column for sorting purposes.