DB Model Changes to Support Admin Features
DB model changes and associated alembic migrations that will support
future admin features.

Adds a `role_id` column with a fk constraint to the `access_keys` table. This
will allow access keys to be assigned to a role in addition to a user, facillitating
access keys that give admin permissions (and potentially other more limited roles, as
needed in the future).

Formalizes the `names` column of the `roles` table as an enumerated type. As we add
roles and enforce permissions, authorization, and access upon them it will be useful
to have a guaranteed set of potential values from which we define permissions.