Add Roles to Existing Permissions Boundaries
Adds roles to existing permissions boundaries in `authentication.py`. Implemented
by creating a wrapper DataClass for `UserData`. Each instance of the `UserData`
class must contain a `User` and a list of active roles. An empty list of active
roles indicates that the user is acting with default permissions. Within
`authorization.py`, we now make use of the `UserRole` Enum. Additional checks
have been included within `permissions.py` that allow users with certain roles
to make additional changes to data models. As new roles are added to MaveDB, they
should be included in the permissions function where appropriate to give users
with those roles additional permissions.

Changes to the routers are to use this DataClass model. In addition:
- You may now create a Access Key with a certain role by calling
`/users/me/access-keys/{role}` with a role that the active user is assigned.
- `get_current_user` has been removed from the `publication_identifiers` routes
as a dependency. This check was not used and is unnecessary.
- The `/users//{id}` PUT route was made more generic to allow for additional
updates to User objects. Some additional Actions were added to `permissions.py`
to support these additional abilities for Admin users.

The User view model was also improved to make use of the UserRole enum. This improves
our ability to validate user provided roles wherever a role is accepted by the API.