Add Support for Active Role Header
Adds support for the `X-Active-Role` header. If no header is
provided, a users' active roles will be all the roles they have
been granted. If the presented active role is `default`, give the
user default user permissions. If the presented active role is something
else, first ensure that the user is a member of that role, and if they
are, allow them to act as it (if they are members of other roles,
they will not have those permissions).

Raise a 401 error if the requested active role is one to which the
user does not belong.