Make Logging Configuration Global
Configures MaveDB logging globally, with several new stock logging configurations that
are initialized on start up. The default config adds a jsonlogger and watchtower handler,
enabling logs to both the console and cloud watch. Note, a filter exists on the cloudwatch
handler so that it only emits canonical logs by default. Additional stock configs for
dev work and debugging were created as well (these do not emit logs to cloudwatch).