Fixed: Executing Jobs Were Sharing DB Sessions
Since the SA session object was defined on worker startup, jobs running at the same time were sharing
a worker session. This could cause unexpected commits and/or rollbacks, in addition to blocking work
in the event that a job left the session in an unrecoverable state. This change creates a session at
the start of each job, defining it on the job's context rather than the worker's context. This will
ensure interactions between jobs and the database are fully independent.