Fixed: Worker Canonical Logs not Output to CloudWatch
Adds the `canonical: True` flag to worker log output. The watchtower handler filters log messages
based on this flag, and does not send them to cloudwatch if it is not present. This change also
changes the log pattern for worker logs, updating them to better take advantage of the JSON log
pattern already used across the API.