Add current column to mapped_variants table
The current column marks whether the mapped variant is considered the most
current mapped version of a given variant in the mapped_variants table.
This is necessary because we may store multiple mapped versions of a variant.
This commit also adds a default value of "1.3" for the VRS version, since
all the mapped variants currently in the database use VRS 1.3, and adds
a default value of the date of execution for the modification date, since
the modification date should keep track of when the row was last updated.