Use Unchanging Identifiers for Worker Jobs
Previously, worker jobs used score set URNs to identify resources on which they should operate.
Due to the nature of these async jobs and the fact that URNs can change on score set publication,
this was a poor choice for processes that might defer execution until a moment in time after the
resource identifier had changed.

This commit updates worker processes to operate on internal database identifiers to identify the
resources on which they should work. This should protect us from instances where the URN changes
and the worker process no longer can access the resource on which it was spawned.