Add users to Collection view model
Only collection admins can see all collection contributors.
Collection viewers and editors can only see a list of collection admins.
The rationale is that collection viewers and editors may want to
contact a collection admin, but do not need to know who other editors
or viewers of the collection are.