Refactor Tests to Better Identify Dependency Separation Issues
Refactors most of the test suite to better identify dependency separation problems. Validation tests
may now be run with only core (and dev) dependencies installed, and fixtures which operate on server dependencies are conditionally loaded based on the installed modules.
With this change, it will be much more straightforward to identify dependency 'leaks', or server dependencies which mistakenly are leaked into validation type code.