Refactor statistics endpoints for simplicity
The previous version of these statistics endpoints featured code that was overly complex for the task at hand. This change refactors statistics code in a way that slightly increases duplicated code (mostly API decorators) while decreasing the complexity of shared functions and logic switches. The result is much clearer and concise endpoints for existing routes.