Simplify Published Variant MV Indexes
Simplifies the indexes on the published variants materialized view to speed up refresh jobs. Previously, the five indexes on this mv created large amounts of write overhead. Due to the nature of concurrent mv refreshes, this caused jobs to run for huge amounts of time with no sign of finishing. The single simpler unique index on only id fields of the table should result in much lower index maintenance overhead and much quicker mv refreshs.