Import all SA models up-front for testing
SA complains about unitialized names for models when some are not imported. This is usually
done in the server_main file, but becomes an issue when testing on only core dependencies
since not all models are loaded and that file is no longer imported by the test suite. Do this
step up front in the top level conftest file instead to ensure we load all model objects into
the SA namespace for testing.