Add OddsPath Property to ScoreRange Data Model
Adds an `OddsPath` property to the score range data model. This property is optional. If it is given, it should contain a `ratios` object with `abnormal` and `normal` (both floats), a `evidenceStrengths` property with `normal` and `abnormal` (both strings), and an optional `source` field.

Uses the PublicationIdentifierBase view model for the OddsPath source. Adds a validator to the score set
record that enforces this publication identifier has been included with the score set as either a primary
or secondary publication identifier. Adds tests for this check and to ensure publications are added successfully.