Add Endpoint for Mapped Variant Fetch by GA4GH Identifier
Given the usefulness of GA4GH VRS Identifiers matching all variants which are given
with respect to the same transcript, it seems prudent to offer an endpoint which allows
users to select mapped variant objects by this identifier. A filter for current variants
is provided, and users must submit a valid GA4GH identifier string. Only variants in
score sets a user has read access to will be exposed.