Fix bugs in mapper job, and centralize some mapper resources
variant_mapper_manager previously failed because of an incorrectly placed
'where' function and because we did two related db queries, one for a score set
and one for target gene(s) within that score set, which resulted in an error
when adding the score set changes to the db. Instead, query to select the score set,
and then loop through the score set's target genes rather than querying the target genes
table in the db.