fix: update gene listing query to fetch distinct HGNC from transcripts
In newer versions of UTA, the gene table has become overrun with mRNA, mitochondrial genes, etc. that polute the result list. Since the real purpose of this endpoint is to list the genes for which we can link transcripts, this query edits the query to select distinct HGNC names from the transcripts table.